Mines no longer blow up when broken in creative

+++ 1.10.2/src/main/java/net/geforcemods/securitycraft/blocks/mines/BlockClaymore.java
@@ -96,7 +96,7 @@ public class BlockClaymore extends BlockContainer implements IExplosive {
 	
     @Override
 	public boolean removedByPlayer(IBlockState state, World world, BlockPos pos, EntityPlayer player, boolean willHarvest){
-    	if (!world.isRemote && !world.getBlockState(pos).getValue(BlockClaymore.DEACTIVATED).booleanValue())
+    	if (!player.capabilities.isCreativeMode && !world.isRemote && !world.getBlockState(pos).getValue(BlockClaymore.DEACTIVATED).booleanValue())
         {
             BlockUtils.destroyBlock(world, pos, false);
             world.createExplosion((Entity) null, (double) pos.getX() + 0.5F, (double) pos.getY() + 0.5F, (double) pos.getZ() + 0.5F, 3.5F, true);
